agentic_ai/requirements.txt
langchain-core>=0.1.17
langgraph>=0.0.19
pydantic>=2.5.2
python-dotenv>=1.0.0
langchain-openai>=0.0.5